reorganized proposal components
