+++ src/components/ui/proposal/ProposalDescription.tsx
